Updating the cluster clone process to only perform copy on initialization in the reactor variables script.

+++ scripts/variables.sh
@@ -37,9 +37,13 @@ if [ ! -z "$ZIMAGI_CLUSTER_DIR" ]; then
     CLUSTER_PROJECT="${__zimagi_lib_dir}/cluster/${ZIMAGI_CLUSTER_DIR}"
     mkdir -p "$CLUSTER_PROJECT"
 
-    for file_path in "${__zimagi_dir}/cluster/"*; do
-        cp -RT "$file_path" "${CLUSTER_PROJECT}/$(basename "$file_path")"
-    done
+    if [ ! -f "${CLUSTER_PROJECT}/.initialized" ]; then
+        for file_path in "${__zimagi_dir}/cluster/"*; do
+            cp -RT "$file_path" "${CLUSTER_PROJECT}/$(basename "$file_path")"
+        done
+        cp -f "${__zimagi_dir}/cluster/.gitignore" "${CLUSTER_PROJECT}/.gitignore"
+        touch "${CLUSTER_PROJECT}/.initialized"
+    fi
 else
     CLUSTER_PROJECT="${__zimagi_dir}/cluster"
 fi
